+++ src/library/renderer/RenderTileMap.h
@@ -11,7 +11,7 @@ class RenderContext;
 class RenderTile;
 class PR_LIB RenderTileMap {
 public:
-	RenderTileMap(size_t xcount, size_t ycount, size_t tilewidth, size_t tileheight);
+	RenderTileMap(uint32 xcount, uint32 ycount, uint32 tilewidth, uint32 tileheight);
 	~RenderTileMap();
 
 	inline size_t tileCount() const { return mTileMap.size(); }
@@ -24,10 +24,10 @@ public:
 	RenderTileStatistics statistics() const;
 
 private:
-	size_t mTileXCount;
-	size_t mTileYCount;
-	size_t mTileWidth;
-	size_t mTileHeight;
+	uint32 mTileXCount;
+	uint32 mTileYCount;
+	uint32 mTileWidth;
+	uint32 mTileHeight;
 	std::vector<RenderTile*> mTileMap;
 };
 }
