+++ src/library/sampler/UniformSampler.cpp
@@ -5,7 +5,7 @@ namespace PR {
 UniformSampler::UniformSampler(uint32 samples)
 	: Sampler()
 	, mSamples(samples)
-	, m2D_X(std::sqrt(samples))
+	, m2D_X(static_cast<uint32>(std::sqrt(samples)))
 	, m2D_Y((mSamples + m2D_X - 1) / m2D_X)
 {
 }
