+++ src/library/spectral/IntensityConverter.cpp
@@ -4,7 +4,7 @@
 namespace PR {
 void IntensityConverter::convert(const Spectrum& s, float& x, float& y, float& z)
 {
-	float avg = s.luminousFlux();
+	float avg = static_cast<float>(s.luminousFlux());
 	x		  = avg;
 	y		  = avg;
 	z		  = avg;
